requirements.txt
# 数据分析Agent依赖
aiohttp>=3.8.0
pydantic>=2.0.0
langfuse>=1.0.0
asyncio-mqtt>=0.11.0
python-dotenv>=1.0.0
openai>=1.0.0
httpx>=0.24.0

# 开发依赖
pytest>=7.0.0
pytest-asyncio>=0.21.0
black>=23.0.0
flake8>=6.0.0